[DPM-108] remove useless use cases
