+++ casino/repository.go
@@ -12,4 +12,6 @@ type Repository interface {
 
 	GetGame(ctx context.Context, gameId uint64) (*models.Game, error)
 	AllGames(ctx context.Context) ([]*models.Game, error)
+
+	GetPlayerInfo(ctx context.Context, accountName string) (*models.PlayerInfo, error)
 }
