+++ server/api/handlers/fetch_casinos.go
@@ -7,7 +7,7 @@ import (
 
 
 func ProcessFetchCasinosRequest(context context.Context, req *interfaces.ApiRequest) (*interfaces.WsResponse, error) {
-	casinos, err := req.UseCases.Casino.AllCasinos(context)
+	casinos, err := req.Repos.Casino.AllCasinos(context)
 	if err != nil {
 		return &interfaces.WsResponse{
 			Type:   "response",
